adding bootstrap4, refactored Homepage, Case, CaseEditor and Organization
