+++ package.json
@@ -21,7 +21,9 @@
     "query-string": "^4.2.3",
     "quill": "github:quilljs/quill#d107d8122363e97f7792bcad040b33a1fbe10cf7",
     "react": "^15.3.2",
+    "react-addons-css-transition-group": "^15.3.2",
     "react-addons-shallow-compare": "^15.3.2",
+    "react-addons-transition-group": "^15.3.2",
     "react-dom": "^15.3.2",
     "react-dropzone-s3-uploader": "github:davidascher/react-dropzone-s3-uploader",
     "react-geosuggest": "^2.0.0",
@@ -33,6 +35,7 @@
     "react-redux": "^4.4.5",
     "react-router": "^2.8.1",
     "react-tap-event-plugin": "^1.0.0",
+    "reactstrap": "^3.8.1",
     "redux": "^3.6.0",
     "redux-form": "^6.1.0",
     "redux-form-material-ui": "^4.1.0",
