+++ src/components/SearchHitCategory/SearchHitCategory.scss
@@ -8,13 +8,15 @@
   color: #ff6f00;
   line-height: 30px;
   padding: 20px 0 14px;
-  border-bottom: 2px solid #eee;
+  // border-bottom: 2px solid #eee;
+  // border-top: 2px solid #eee;
 }
 
 .results {
-  padding: 32px 0 0;
+  // padding: 32px 0 0;
   overflow: hidden;
-  border-bottom: 2px solid #eee;
+  // border-bottom: 2px solid #eee;
+  // border-bottom: 2px solid #eee;
   display: flex;
   flex-wrap: wrap;
 }
