+++ src/components/SubAccordionTab.scss
@@ -1,6 +1,7 @@
 @import "../UniversalStyles";
 
 @mixin tab {
+  border-bottom: 1px solid #ccc;
   padding: 16px;
   padding-right: 96px;
   position: relative;
@@ -61,6 +62,10 @@
   }
 }
 
+.subtab {
+  @include tab;
+}
+
 .tab-open {
   @include tab;
 
@@ -72,3 +77,5 @@
     display: block;
   }
 }
+
+
